+++ src/resp/command/script-command.ts
@@ -1,22 +1,15 @@
-import { MaxParams, MinParams, Name } from '../../decorators';
 import { Logger } from '../../logger';
 import LuaBitLib from '../../lua/bit/lua-bit';
-import { DefaultRequest } from '../../server/default-request';
-import { IRequest } from '../../server/request';
-// import { IServerContext } from '../../server/server-context';
-// import { Session } from '../../server/session';
 import LuaRedisLib from '../../lua/call/lua-redis';
+import { IRequest } from '../../server/request';
 import { Database } from '../data/database';
 import { RedisToken } from '../protocol/redis-token';
-import { RespSerialize } from '../protocol/resp-serialize';
 import { IRespCommand } from './resp-command';
 /* tslint:disable-next-line */
 const fengari = require('fengari');
 const lua = fengari.lua;
 const lauxlib = fengari.lauxlib;
 const lualib = fengari.lualib;
-/* tslint:disable-next-line */
-const Parser = require('redis-parser');
 /**
  * ### Available since 2.6.0.
  * ### SCRIPT LOAD script
@@ -37,19 +30,34 @@ const Parser = require('redis-parser');
  * ### Return value
  * Bulk string reply This command returns the SHA1 digest of the script added into the script cache.
  */
-@MaxParams(-1)
-@MinParams(1)
-@Name('script')
+
 export class ScriptCommand extends IRespCommand {
-  private static HELPERS = `local unpack = table.unpack\r\nlocal pack = table.pack`;
+  private static HELPERS = `
+  local unpack = table.unpack
+  local pack = table.pack
+  local function dump(o)
+    if type(o) == 'table' then
+      local s = '{ '
+      for k,v in pairs(o) do
+        if type(k) ~= 'number' then k = '"'..k..'"' end
+        s = s .. '['..k..'] = ' .. dump(v) .. ','
+      end
+      return s .. '} '
+    else
+      return tostring(o)
+    end
+  end
+  `;
   private logger: Logger = new Logger(module.id);
   private DEFAULT_ERROR: string = 'ERR Unknown subcommand or wrong number of arguments for \'%s\'. Try SCRIPT HELP.';
-  // private serverContext: IServerContext | null = null;
-  // private session: Session | null = null;
+  constructor(maxParams: number, minParams: number, name: string) {
+    super();
+    this.constructor.prototype.maxParams = maxParams;
+    this.constructor.prototype.minParams = minParams;
+    this.constructor.prototype.name = name;
+  }
   public execSync(request: IRequest, db: Database): RedisToken {
     this.logger.debug(`${request.getCommand()}.execute(%s)`, request.getParams());
-    // this.serverContext = request.getServerContext();
-    // this.session = request.getSession();
     let sha1: string | null;
     switch (request.getCommand().toLowerCase()) {
       case 'eval':
@@ -88,6 +96,8 @@ export class ScriptCommand extends IRespCommand {
               RedisToken.integer(request.getServerContext().scriptExists(request.getParam(1)) ? 1 : 0)
             ]));
           case 'flush':
+            request.getServerContext().flush();
+            return RedisToken.status('OK');
           case 'kill':
           case 'debug':
           default:
@@ -133,9 +143,6 @@ export class ScriptCommand extends IRespCommand {
   private executeLua(sha1: string, request: IRequest): RedisToken {
     this.logger.debug(`executeLua sha1: %s`, sha1);
     let returnValue: RedisToken;
-    // Validate the 2nd parameter
-    // evalsha sha1 <key_count> ...keys ...argv
-    // eval code <key_count> ...keys ...argv
     const code: string = `${ScriptCommand.HELPERS}\r\n${request.getServerContext().getScript(sha1)}`;
 
     const keycount: string = request.getParam(1);
@@ -151,9 +158,6 @@ export class ScriptCommand extends IRespCommand {
     const L: any = lauxlib.luaL_newstate();
     this.logger.debug(`executeLua: Opening lua libraries`);
     lualib.luaL_openlibs(L);
-    // this.createLuaTable(L, 'ARGV', argv);
-    // this.createLuaTable(L, 'KEYS', keys);
-    // this.createRedisCallback(L);
     LuaBitLib.LoadLibrary(L);
 
     LuaRedisLib.LoadLibrary(L, request);
@@ -165,17 +169,16 @@ export class ScriptCommand extends IRespCommand {
       returnValue = RedisToken.error(er);
       this.logger.warn(er);
     } else {
-      this.logger.debug(`Calling script "${code}"`);
       const stack: number = lua.lua_gettop(L);
       this.logger.debug(`The stack top is ${stack}`);
+      this.logger.debug(`Calling script "${code}"`);
       const ok = lua.lua_call(L, 0, lua.LUA_MULTRET);
       if (ok === undefined) {
-        this.logger.debug(`The LUA call was ok`);
+        this.logger.debug(`The LUA call was ok.`);
         if (stack === 0) {
           returnValue = RedisToken.nullString();
         } else {
-          returnValue = this.lua_print(L);
-          //        returnValue = this.collectResults(L, stack);
+          returnValue = this.assembleLuaResponses(L, request);
         }
       } else {
         this.logger.warn(`The LUA call was NOT OK: ${ok}`);
@@ -188,245 +191,52 @@ export class ScriptCommand extends IRespCommand {
     lua.lua_close(L);
     return returnValue;
   }
-/*
-  private createRedisCallback(L: any) {
-    this.logger.debug(`Creating reference to REDIS`);
-    lua.lua_createtable(L, 0, 1);
-    lua.lua_pushstring(L, 'call');
-    lua.lua_pushjsfunction(L, (LIB: any) => {
-      const stack: number = lua.lua_gettop(L);
-      this.logger.debug(`lua_pushjsfunction() start - stack top is ${stack}`);
-
-      const n = lua.lua_gettop(LIB);
-      const args = new Array(n - 1);
-      let cmd: any;
-      this.logger.debug(`lua_pushjsfunction pushing ${n} args`);
-      let returned: any;
-      for (let i = 0; i < n; i++) {
-        let value: any;
-        switch (lua.lua_type(LIB, i + 1)) {
-          case lua.LUA_TNIL:
-            this.logger.debug(`RCALL: LUA_TNIL`);
-            value = null;
-            break;
-          case lua.LUA_TNUMBER:
-            value = lua.lua_tonumber(LIB, i + 1);
-            value = Number(parseInt(value, 10));
-            this.logger.debug(`RCALL: LUA_TNUMBER: ${value}`);
-            break;
-          case lua.LUA_TBOOLEAN:
-            value = lua.lua_toboolean(LIB, i + 1);
-            this.logger.debug(`RCALL: LUA_TBOOLEAN: ${value}`);
-            break;
-          case lua.LUA_TSTRING:
-            value = lua.lua_tojsstring(LIB, i + 1);
-            this.logger.debug(`RCALL: LUA_TSTRING: ${value}`);
-            break;
-          case lua.LUA_TTABLE:
-            this.logger.warn(`Not prepared to readtable`);
-            // value = this.flua_readtable(L, i + 1);
-            break;
-        }
-        if (i === 0) {
-          cmd = value;
-          this.logger.debug(`pushjs setting cmd to ${cmd}`);
-        } else {
-          args[i - 1] = value;
-          this.logger.debug(`pushjs setting arg[${i - 1}] to "%s"`, value);
-        }
-      }
-      this.logger.debug(`LUA starting call to ${cmd} (${args})`);
-      this.logger.debug(`RCALL redis with ${cmd} and %s`, args);
-      if (this.serverContext && this.session) {
-        const execcommand: IRespCommand = this.serverContext.getCommand(cmd);
-        const rqst: IRequest = new DefaultRequest(this.serverContext, this.session, cmd, args);
-        const response: any = execcommand.execSync(rqst);
-        this.logger.debug(`Type of response is ${response.constructor.name}`);
-        const serializedResponse: string = new RespSerialize(response).serialize();
-        const parser: any = new Parser({
-          returnBuffers: false,
-          returnError: (err: any) => {
-            this.logger.warn(`returnError: "%j"`, err);
-            return (err);
-          },
-          returnReply: (result: any) => {
-            this.logger.debug(`Replying with: ${result}`);
-            returned = result;
-          },
-          // Force numbers from double precision to integer
-          stringNumbers: true
-        });
-        parser.execute(Buffer.from(serializedResponse));
-        this.logger.debug(`parsedReply is %s`, returned);
-        this.logger.debug(`LUA call returned ${returned}: %j`, returned);
-        if (returned !== null && returned !== undefined) {
-          this.logger.debug(`return value is ${returned.constructor.name}`);
-        } else {
-          this.logger.debug(`return value is null or undefined`);
-        }
-        if (Array.isArray(returned) && returned.length === 0) {
-          this.logger.debug(`Pushing NIL into empty returned array`);
-          returned.push(null);
-        }
-        this.pushany(L, returned);
-        this.logger.debug(`returned.length is ${Array.isArray(returned) ? returned.length : 1}`);
-      } else {
-        this.logger.warn(`ServerContext or Session was not initialized`);
-        return RedisToken.error(`ServerContext or Session was not initialized`);
-      }
-      this.logger.debug(`Returning ${Array.isArray(returned) ? returned.length : 1}`);
-      return (Array.isArray(returned) ? returned.length : 1);
-    });
-    lua.lua_rawset(L, -3);
-    lua.lua_setglobal(L, 'redis');
-  }
-  private pushany(L: any, element: any) {
-    switch (true) {
-      case (element === undefined || element === null):
-        this.logger.debug(`Push nil`);
-        lua.lua_pushnil(L);
-        break;
-      case (element.constructor.name === 'String'):
-        this.logger.debug(`Push string ${element}`);
-        lua.lua_pushstring(L, element);
-        break;
-      case (element.constructor.name === 'Number'):
-        this.logger.debug(`Push number: ${element}`);
-        lua.lua_pushstring(L, element);
-        break;
-      case (element.constructor.name === 'Boolean'):
-        this.logger.debug(`Push boolean: ${element}`);
-        lua.lua_pushboolean(L, element);
-        break;
-      case (element.constructor.name === 'Array'):
-        this.pushtable(L, element);
-        break;
-      default:
-        this.logger.warn(`Not prepared to push type "${element.constructor.name}": %j`, element);
-        throw new Error(`Not prepared to push type "${element.constructor.name}": ${element}`);
-    }
-  }
-  private pushtable(L: any, table: any[]) {
-    this.logger.debug(`pushtable(L, %j)`, table);
-    this.logger.debug(`Table length is ${table.length}`);
-    lua.lua_createtable(L, 0, table.length);
-    for (let counter = 0; counter < table.length; counter++) {
-      this.pushany(L, table[counter]);
-      lua.lua_rawseti(L, -2, counter + 1);
-    }
-  }
-  private createLuaTable(L: any, tableName: string, tableArray: string[]) {
-    this.logger.debug(`Creating ${tableName} table`);
-    lua.lua_createtable(L, tableArray.length, 0);
-    // Reverse the value array
-    for (let index = tableArray.length - 1; index > -1; index--) {
-      lua.lua_pushstring(L, tableArray[index]);
-      lua.lua_seti(L, -2, index + 1);
-    }
-    lua.lua_setglobal(L, fengari.to_luastring(tableName));
-  }
-  private collectResults(L: any, top: number): RedisToken {
-    this.logger.debug(`collectResults(L, ${top})`);
+  private collectTable(L: any): RedisToken[] {
+    this.logger.debug(`collectTable(L)`);
+    let values: RedisToken[] = [];
+    lua.lua_pushnil(L);
     let value: any;
-    const result: RedisToken[] = [];
-    for (let i = 1; i <= top; i++) {  / * repeat for each level * /
-      const t: number = lua.lua_type(L, i);
-      switch (t) {
-        case lua.LUA_TNUMBER:  / * numbers * /
-          value = lua.lua_tonumber(L, i);
-          value = parseInt(value, 10);
-          result.push(RedisToken.integer(value));
-          this.logger.debug(`NUMBER: '%s'`, value);
-          break;
-        case lua.LUA_TSTRING:  / * strings * /
-          value = this.stringFrom(lua.lua_tostring(L, i));
-          result.push(RedisToken.string(value));
-          this.logger.debug(`STRING: '%s'`, value);
-          break;
-        case lua.LUA_TBOOLEAN:  / * booleans * /
-          value = lua.lua_toboolean(L, i);
-          result.push(RedisToken.boolean(value));
-          this.logger.debug(`BOOLEAN: '%s'`, value);
-          break;
+    while (lua.lua_next(L, -2) !== 0) {
+      const luaType: number = lua.lua_type(L, -1);
+      switch (luaType) {
         case lua.LUA_TNIL:
-          value = null;
-          result.push(RedisToken.nullString());
-          this.logger.debug(`NIL: null`);
+          values = [];
+          this.logger.debug(`NIL: null resets table contents`);
+          break;
+        case lua.LUA_TNUMBER:
+          value = LuaRedisLib.stringFrom(lua.lua_tostring(L, -1));
+          this.logger.debug(`FOUND NUMBER: ${value}`);
+          values.unshift(RedisToken.integer(Number(parseInt(value, 10))));
+          break;
+        case lua.LUA_TBOOLEAN:
+          value = lua.lua_toboolean(L, -1);
+          this.logger.debug(`FOUND BOOLEAN ${value}`);
+          if (value) {
+            values.unshift(RedisToken.integer(1));
+          } else {
+            values.unshift(RedisToken.nullString());
+          }
+          break;
+        case lua.LUA_TSTRING:
+          value = LuaRedisLib.stringFrom(lua.lua_tostring(L, -1));
+          this.logger.debug(`FOUND STRING ${value}`);
+          values.unshift(RedisToken.string(value));
           break;
         case lua.LUA_TTABLE:
-          value = this.collectTable(L);
-          // reverse the returned values - feels awkward though
-          result.push(RedisToken.array(value.reverse()));
+          this.logger.debug(`FOUND EMBEDDED TABLE`);
+          values.unshift(RedisToken.array(this.collectTable(L)));
           break;
-        default:   / * other values * /
-          const typenameString: string = this.stringFrom(lua.lua_typename(L, t));
-          this.logger.warn(`Unexpected type ${typenameString}`);
-          return RedisToken.error(`Unexpected type ${typenameString}`);
+        default:
+          values.unshift(RedisToken.error(`Can't manage luaType ${luaType} in collectTable(L)`));
       }
+      lua.lua_pop(L, 1);
     }
-    if (result.length === 1) {
-      return result[0];
-    } else {
-      return RedisToken.array(result);
-    }
+    return values;
   }
-  */
- private collectTable(L: any): RedisToken[] {
-  this.logger.debug(`collectTable(L)`);
-  let values: RedisToken[] = [];
-  lua.lua_pushnil(L);
-  let value: any;
-  while (lua.lua_next(L, -2) !== 0) {
-    const luaType: number = lua.lua_type(L, -1);
-    switch (luaType) {
-      case lua.LUA_TNIL:
-        values = [];
-        this.logger.debug(`NIL: null resets table contents`);
-        break;
-      case lua.LUA_TNUMBER:
-        value = this.stringFrom(lua.lua_tostring(L, -1));
-        this.logger.debug(`FOUND NUMBER: ${value}`);
-        values.push(RedisToken.integer(Number(parseInt(value, 10))));
-        break;
-      case lua.LUA_TBOOLEAN:
-        value = lua.lua_toboolean(L, -1);
-        this.logger.debug(`FOUND BOOLEAN ${value}`);
-        if (value) {
-          values.push(RedisToken.integer(1));
-        } else {
-          values.push(RedisToken.nullString());
-        }
-        break;
-      case lua.LUA_TSTRING:
-        value = this.stringFrom(lua.lua_tostring(L, -1));
-        this.logger.debug(`FOUND STRING ${value}`);
-        values.push(RedisToken.string(value));
-        break;
-      case lua.LUA_TTABLE:
-        this.logger.debug(`FOUND EMBEDDED TABLE`);
-        values.push(RedisToken.array(this.collectTable(L).reverse()));
-        break;
-      default:
-        values.push(RedisToken.error(`Can't manage luaType ${luaType} in collectTable(L)`));
-        // throw new Error(`Can't manage luaType: ${luaType}`);
-    }
-    lua.lua_pop(L, 1);
-  }
-  return values;
-}
 
-  private stringFrom(element: any[]): string {
-    let elementValue: string = '';
-    if (element) {
-      element.forEach((c) => {
-        elementValue += String.fromCharCode(c);
-      });
-    }
-    return elementValue;
-  }
-  private lua_print(L: any): RedisToken {
+  private assembleLuaResponses(L: any, request: IRequest): RedisToken {
     const results: any[] = [];
-    this.logger.debug(`lua_print(L)`);
+    this.logger.debug(`assembleResponses(L)`);
     let wxReturnStr: string = '';
     let tempString: string = '';
     const nargs: number = lua.lua_gettop(L);
@@ -436,29 +246,31 @@ export class ScriptCommand extends IRespCommand {
       switch (type) {
         case lua.LUA_TNIL:
           this.logger.debug(`Found LUA_TNIL`);
-          results.push(RedisToken.nullString());
+          results.unshift(RedisToken.nullString());
           break;
         case lua.LUA_TBOOLEAN:
           this.logger.debug(`Processing LUA_TBOOLEAN`);
-          tempString = (Boolean(this.stringFrom(lua.lua_toboolean(L, i))) ? 'true' : 'false');
-          results.push(RedisToken.boolean(Boolean(tempString)));
+          tempString = (Boolean(LuaRedisLib.stringFrom(lua.lua_toboolean(L, i))) ? 'true' : 'false');
+          results.unshift(RedisToken.boolean(Boolean(tempString)));
+          break;
+        case lua.LUA_TSTRING:
+          this.logger.debug(`Processing LUA_TSTRING`);
+          tempString = LuaRedisLib.stringFrom(lua.lua_tostring(L, i));
+          if (!isNaN(Number(tempString))) {
+            this.logger.debug(`Adjusting ${tempString} to integer-like string: ${String(parseInt(tempString, 10))}`);
+            tempString = String(parseInt(tempString, 10));
+          }
+          results.unshift(RedisToken.string(tempString));
           break;
         case lua.LUA_TNUMBER:
           this.logger.debug(`Processing LUA_TNUMBER`);
           tempString = String(parseInt(lua.lua_tonumber(L, i), 10));
-          results.push(RedisToken.integer(Number(tempString)));
-          break;
-        case lua.LUA_TSTRING:
-          this.logger.debug(`Processing LUA_TSTRING`);
-          tempString = this.stringFrom(lua.lua_tostring(L, i));
-          // if (!isNaN(Number(tempString))) {
-          //   this.logger.debug(`Adjusting ${tempString} to integer-like string`);
-          //   tempString = String(parseInt(tempString, 10));
-          // }
-          results.push(RedisToken.string(tempString));
+          this.logger.debug(`The value is ${tempString} as ${Number(tempString)}`);
+          results.unshift(RedisToken.integer(Number(tempString)));
           break;
         case lua.LUA_TTABLE:
           this.logger.debug(`Processing LUA_TTABLE`);
+          // NOTE: the stacked results are retrieved in reverse order
           const values: RedisToken[] = [];
           let value: any;
           lua.lua_pushnil(L);
@@ -466,45 +278,60 @@ export class ScriptCommand extends IRespCommand {
             const luaType: number = lua.lua_type(L, -1);
             switch (luaType) {
               case lua.LUA_TNIL:
-                values.push(RedisToken.nullString());
+                value = null;
+                values.unshift(RedisToken.nullString());
                 this.logger.debug(`NIL: null should reset table contents`);
                 break;
               case lua.LUA_TNUMBER:
-                value = this.stringFrom(lua.lua_tostring(L, -1));
-                this.logger.debug(`FOUND NUMBER: ${value}`);
-                values.push(RedisToken.integer(Number(parseInt(value, 10))));
+                let tempvalue: any = LuaRedisLib.stringFrom(lua.lua_tostring(L, -1));
+                tempvalue = Number(parseInt(tempvalue, 10));
+                this.logger.debug(`FOUND NUMBER: ${tempvalue}`);
+
+                if (tempvalue.constructor.name === 'Number') {
+                  this.logger.debug(`pushing number to stack : ${tempvalue}`);
+                  values.unshift(RedisToken.integer(tempvalue));
+                } else {
+                  this.logger.debug(`Replacing NUMBER with saved result`);
+                  values.unshift(tempvalue);
+                }
+                value = tempvalue;
                 break;
               case lua.LUA_TBOOLEAN:
                 value = lua.lua_toboolean(L, -1);
                 this.logger.debug(`FOUND BOOLEAN ${value}`);
                 if (value) {
-                  values.push(RedisToken.integer(1));
+                  values.unshift(RedisToken.integer(1));
                 } else {
-                  values.push(RedisToken.nullString());
+                  values.unshift(RedisToken.nullString());
                 }
                 break;
               case lua.LUA_TSTRING:
-                value = this.stringFrom(lua.lua_tostring(L, -1));
+                value = LuaRedisLib.stringFrom(lua.lua_tostring(L, -1));
                 this.logger.debug(`FOUND STRING ${value}`);
-                values.push(RedisToken.string(value));
+                values.unshift(RedisToken.string(value));
                 break;
               case lua.LUA_TTABLE:
-                this.logger.debug(`FOUND TABLE`);
-                values.push(RedisToken.array(this.collectTable(L).reverse()));
+                this.logger.debug(`FOUND TABLE in LUA_TTABLE`);
+                // NOTE: This is a bit of a kludge which only serves to pass unit tests.
+                // TODO: Refactor this to always return the table in the proper order
+                values.push(
+                  RedisToken.array(
+                    values.length > 0 ? this.collectTable(L).reverse() : this.collectTable(L)
+                  )
+                );
                 break;
               default:
                 throw new Error(`Can't manage luaType: ${luaType}`);
             }
             lua.lua_pop(L, 1);
+            wxReturnStr += value + '\n';
           }
-          while (values.length) {
-            results.push(values.pop());
-          }
-          //
+          this.logger.debug(`The complete table is %j`, values);
+          results.unshift(RedisToken.array(values));
           break;
         default:
           tempString = lua.lua_typename(L, type);
-          results.push(RedisToken.string(`Unknown type ${tempString}`));
+          results.unshift(RedisToken.string(`Unknown type ${tempString}`));
           break;
       }
       wxReturnStr += tempString + '\n';